+++ CHANGELOG.md
@@ -5,10 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Changed
+- Renamed CLI args to `test_file_pattern`, `test_name_pattern`, `hide_passed_tests`, `print_results`, `results_file`, and `fail_fast`.
+
 ## [1.2.0]
 
 ### Added
- - `stop_on_first_failed_test` CLI arg
+- `stop_on_first_failed_test` CLI arg
 
 ## [1.1.0]